pyproject.toml
[project]
name = "commspt-nonebot-plugin"
dynamic = ["version"]
description = "For LittleSkin Commspt."
authors = [{ name = "FalfaChino", email = "falfa@chiyoda.moe" }]
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-plugin-alconna>=0.43.0",
    "httpx>=0.24",
    "jinja2>=3.1.2",
    "yggdrasil-mc>=0.2.3",
    "arrow>=1.3.0",
    "email-validator>=2.1.1",
    "pytz>=2024.1",
    "pillow>=10.3.0",
]
requires-python = ">=3.10,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/FalfaChino/commspt-nonebot-plugin"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "commspt_nonebot_plugin/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
